+++ src/client/actions/districtDrawing.ts
@@ -42,6 +42,8 @@ export const setGeoLevelVisibility = createAction("Set geolevel visibility")<rea
 
 export const toggleDistrictLocked = createAction("Toggle district locked")<DistrictId>();
 
+export const toggleLimitDrawingToWithinCounty = createAction("Limit drawing to within county")();
+
 export const showAdvancedEditingModal = createAction("Show advanced editing warning modal")<
   boolean
 >();