+++ src/client/components/map/Map.tsx
@@ -72,6 +72,7 @@ interface Props {
   readonly evaluateMetric?: EvaluateMetric;
   readonly evaluateMode: boolean;
   readonly isReadOnly: boolean;
+  readonly limitSelectionToCounty: boolean;
   readonly findMenuOpen: boolean;
   readonly findTool: FindTool;
   readonly label?: string;
@@ -131,6 +132,7 @@ const DistrictsMap = ({
   geoLevelIndex,
   lockedDistricts,
   isReadOnly,
+  limitSelectionToCounty,
   findMenuOpen,
   evaluateMetric,
   evaluateMode,
@@ -224,6 +226,7 @@ const DistrictsMap = ({
     // eslint-disable-next-line
   }, [mapRef]);
 
+  // Keyboard handlers
   function downHandler({ key }: KeyboardEvent) {
     if (key === "Spacebar" || key === " ") {
       setTogglePan(true);
@@ -235,6 +238,7 @@ const DistrictsMap = ({
       setTogglePan(false);
     }
   };
+
   // Add event listeners
   useEffect(() => {
     window.addEventListener("keydown", downHandler);
@@ -578,6 +582,7 @@ const DistrictsMap = ({
             project.districtsDefinition,
             lockedDistricts,
             staticGeoLevels,
+            limitSelectionToCounty,
             setSelectionInProgress
           );
         } else if (selectionTool === SelectionTool.PaintBrush) {
@@ -588,7 +593,9 @@ const DistrictsMap = ({
             project.districtsDefinition,
             lockedDistricts,
             staticGeoLevels,
-            setSelectionInProgress
+            setSelectionInProgress,
+            limitSelectionToCounty,
+            selectedGeounits
           );
         }
       }
@@ -604,7 +611,8 @@ const DistrictsMap = ({
     lockedDistricts,
     isReadOnly,
     evaluateMode,
-    isPanning
+    isPanning,
+    limitSelectionToCounty
   ]);
 
   return (