+++ src/client/reducers/districtDrawing.ts
@@ -28,7 +28,8 @@ import {
   saveDistrictsDefinition,
   setSavingState,
   FindTool,
-  selectEvaluationMetric
+  selectEvaluationMetric,
+  toggleLimitDrawingToWithinCounty
 } from "../actions/districtDrawing";
 import { updateDistrictsDefinition, updateDistrictLocks } from "../actions/projectData";
 import { SelectionTool } from "../actions/districtDrawing";
@@ -110,6 +111,7 @@ export interface DistrictDrawingState {
   readonly evaluateMetric: EvaluateMetric | undefined;
   readonly findIndex?: number;
   readonly findTool: FindTool;
+  readonly limitSelectionToCounty: boolean;
   readonly saving: SavingState;
   readonly undoHistory: UndoHistory;
 }
@@ -124,6 +126,7 @@ export const initialDistrictDrawingState: DistrictDrawingState = {
   evaluateMode: false,
   evaluateMetric: undefined,
   findTool: FindTool.Unassigned,
+  limitSelectionToCounty: false,
   saving: "unsaved",
   undoHistory: {
     past: [],
@@ -205,6 +208,11 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
       return {
         ...state,
         saving: action.payload
+      };
+    case getType(toggleLimitDrawingToWithinCounty):
+      return {
+        ...state,
+        limitSelectionToCounty: !state.limitSelectionToCounty
       };
     case getType(setHighlightedGeounits):
       return {